requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
outcome==1.3.0.post0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
PySocks==1.7.1
python-dotenv==1.0.1
redis==5.2.1
selenium==4.27.1
service-identity==24.2.0
setuptools==75.6.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
